# mimic-harness/contracts/openapi/agent-api.yaml
openapi: 3.1.0
info:
  title: Ninic Creative Agent Studio Agent API
  version: 0.1.0
  description: >
    Phase 1 正式 Contracts 基线。前端、Java Runtime、Python Runtime 必须共同遵守本文件。
    本阶段只覆盖 Mock SSE 主链路，不接真实模型、数据库、RAG、MQ、MCP 或对象存储。
servers:
  - url: http://localhost:8080
    description: Java Runtime
  - url: http://localhost:8000
    description: Python Runtime
...
paths:
...
  /api/threads:
    get:
      tags: [Threads]
      summary: 会话列表
      operationId: listThreads
      responses:
        "200":
          description: 会话列表
          content:
            application/json:
              schema:
                type: object
                required: [threads]
                properties:
                  threads:
                    type: array
                    items:
                      $ref: "#/components/schemas/Thread"
    post:
      tags: [Threads]
      summary: 创建会话
      operationId: createThread
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThreadRequest"
      responses:
        "200":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
  /api/threads/{threadId}:
    get:
      tags: [Threads]
      summary: 会话详情
      operationId: getThread
      parameters:
        - $ref: "#/components/parameters/ThreadId"
      responses:
        "200":
          description: 会话详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/threads/{threadId}/messages:
    get:
      tags: [Threads]
      summary: 会话消息
      operationId: listMessages
      parameters:
        - $ref: "#/components/parameters/ThreadId"
      responses:
        "200":
          description: 消息列表
          content:
            application/json:
              schema:
                type: object
                required: [messages]
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
  /api/threads/{threadId}/runs/stream:
    post:
      tags: [Runs]
      summary: 创建 AIGC Campaign Run 并通过 SSE 返回流式事件
      operationId: streamCampaignRun
      parameters:
        - $ref: "#/components/parameters/ThreadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequest"
      responses:
        "200":
          description: SSE 流式响应
          content:
            text/event-stream:
              schema:
                type: string
                description: 事件格式见 contracts/events/sse-events.md
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/runs/{runId}:
    get:
      tags: [Runs]
      summary: Run 详情
      operationId: getRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Run 详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    ThreadId:
      name: threadId
      in: path
      required: true
      schema:
        type: string
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    RuntimeType:
      type: string
      enum: [java, python]
...
    CreateThreadRequest:
      type: object
      properties:
        title:
          type: string
          examples: [夏季气泡咖啡营销素材生成]
        campaignId:
          type: string
    Thread:
      type: object
      required: [threadId, title, createdAt, updatedAt]
      properties:
        threadId:
          type: string
        campaignId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Message:
      type: object
      required: [messageId, threadId, role, content, createdAt]
      properties:
        messageId:
          type: string
        threadId:
          type: string
        role:
          type: string
          enum: [user, assistant, system, tool]
        content:
          type: string
        runId:
          type: string
        createdAt:
          type: string
          format: date-time
    RunStatus:
      type: string
      enum: [created, running, succeeded, failed, cancelled]
    RunScenario:
      type: string
      enum: [summer_bubble_coffee_campaign]
    RunRequest:
      type: object
      required: [prompt, runtime, scenario, modelProvider, model, mockMode]
      properties:
        prompt:
          type: string
        runtime:
          $ref: "#/components/schemas/RuntimeType"
        scenario:
          $ref: "#/components/schemas/RunScenario"
        modelProvider:
          type: string
          examples: [mock]
        model:
          type: string
          examples: [mock-creative-agent-v1]
        mockMode:
          type: boolean
    Run:
      type: object
      required: [runId, threadId, runtime, scenario, status, createdAt]
      properties:
        runId:
          type: string
        threadId:
          type: string
        runtime:
          $ref: "#/components/schemas/RuntimeType"
        scenario:
          $ref: "#/components/schemas/RunScenario"
        status:
          $ref: "#/components/schemas/RunStatus"
        modelProvider:
          type: string
        model:
          type: string
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        errorCode:
          type: string
        errorMessage:
          type: string
...
    ErrorResponse:
      type: object
      required: [code, message, createdAt]
      properties:
        code:
          type: string
          enum:
            - RUNTIME_UNAVAILABLE
            - SSE_CONNECT_FAILED
            - SSE_STREAM_INTERRUPTED
            - RUN_FAILED
            - INVALID_REQUEST
            - INVALID_RUNTIME
            - UNKNOWN_EVENT_TYPE
            - INVALID_ASSET_PAYLOAD
            - EMPTY_HTML_PREVIEW
            - NOT_FOUND
        message:
          type: string
        detail:
          type: string
        requestId:
          type: string
        runId:
          type: string
        runtime:
          $ref: "#/components/schemas/RuntimeType"
        createdAt:
          type: string
          format: date-time